Fix blocked commands issue.
This commit fixes the blocked commands bug, where the command would be blocked, but when using more than 1 argument (more than the actual blocked command, it wouldn't block the command being used. Example: If you type /home (as an example of a blocked command), it would say it's blocked, yes, that works fine, but whenever you type like: /home test, it wouldn't block the command, thus meaning you were able to bypass the system. Now, this fixes the bypass, as tested.